mcp-servers/se-demo/requirements.txt
fastmcp>=2.0.0
mcp>=1.0.0
snowflake-connector-python>=3.12.0
httpx>=0.26.0
requests>=2.32.0
python-dotenv>=1.0.0
cryptography>=42.0.0
